break up baseline module
